+++ tests/test_cli.py
@@ -250,6 +250,74 @@ from tests.utils import path_to_test_resource
                 "langid_expected_languages": "de,es,cs",
             },
         ),
+        (
+            ["batch", "transcribe", "--summarize"],
+            {
+                "summarize": True,
+            },
+        ),
+        (
+            [
+                "batch",
+                "transcribe",
+                "--summarize",
+                "--summary-content-type=informative",
+            ],
+            {
+                "summarize": True,
+                "content_type": "informative",
+            },
+        ),
+        (
+            [
+                "batch",
+                "transcribe",
+                "--summarize",
+                "--summary-content-type=informative",
+            ],
+            {
+                "summarize": True,
+                "content_type": "informative",
+            },
+        ),
+        (
+            ["batch", "transcribe", "--summarize", "--summary-length=detailed"],
+            {
+                "summarize": True,
+                "summary_length": "detailed",
+            },
+        ),
+        (
+            ["batch", "transcribe", "--summarize", "--summary-type=paragraphs"],
+            {
+                "summarize": True,
+                "summary_type": "paragraphs",
+            },
+        ),
+        (
+            ["batch", "transcribe", "--sentiment-analysis"],
+            {
+                "sentiment_analysis": True,
+            },
+        ),
+        (
+            ["batch", "transcribe", "--detect-topics"],
+            {
+                "detect_topics": True,
+            },
+        ),
+        (
+            [
+                "batch",
+                "transcribe",
+                "--detect-topics",
+                "--topics=topic1,topic2,topic3",
+            ],
+            {
+                "detect_topics": True,
+                "topics": "topic1,topic2,topic3",
+            },
+        ),
     ],
 )
 def test_cli_arg_parse_with_file(args, values):
@@ -622,6 +690,9 @@ def test_rt_main_with_config_file(mock_server):
     assert msg["transcription_config"]["domain"] == "fake"
     assert msg["transcription_config"]["enable_entities"] is True
     assert msg["transcription_config"].get("operating_point") is None
+    assert msg["translation_config"] is not None
+    assert msg["translation_config"]["enable_partials"] is False
+    assert msg["translation_config"]["target_languages"] == ["es"]
 
 
 def test_rt_main_with_config_file_cmdline_override(mock_server):
@@ -636,6 +707,8 @@ def test_rt_main_with_config_file_cmdline_override(mock_server):
         mock_server.url,
         "--config-file",
         config_path,
+        "--translation-langs=fr",
+        "--enable-translation-partials",
         "--auth-token=xyz",
         "--lang=yz",
         "--output-locale=en-US",
@@ -665,6 +738,9 @@ def test_rt_main_with_config_file_cmdline_override(mock_server):
     assert msg["transcription_config"]["output_locale"] == "en-US"
     assert msg["transcription_config"]["speaker_change_sensitivity"] == 0.8
     assert msg["transcription_config"]["operating_point"] == "enhanced"
+    assert msg["translation_config"] is not None
+    assert msg["translation_config"]["enable_partials"] is True
+    assert msg["translation_config"]["target_languages"] == ["fr"]
 
 
 @pytest.mark.parametrize(