+++ tests/test_models.py
@@ -1,4 +1,6 @@
-import pytest
+from dataclasses import asdict
+
+from pytest import mark, param
 
 from speechmatics import models
 from speechmatics.batch_client import BatchClient
@@ -53,7 +55,7 @@ def test_translationconfig_default_values():
     assert {"target_languages": None, "enable_partials": False} == config.asdict()
 
 
-@pytest.mark.parametrize(
+@mark.parametrize(
     "target_languages, enable_partials, want_config",
     [
         (["fr"], False, {"target_languages": ["fr"], "enable_partials": False}),
@@ -72,7 +74,7 @@ def test_translationconfig(target_languages, enable_partials, want_config):
     assert want_config == config.asdict()
 
 
-@pytest.mark.parametrize(
+@mark.parametrize(
     "url, want",
     [
         ("example.com/v2", "example.com/v2"),
@@ -86,3 +88,115 @@ def test_connection_settings_url(url, want):
     batch_client = BatchClient(connection_settings)
     got = batch_client.connection_settings.url
     assert got == want
+
+
+@mark.parametrize(
+    "params, want",
+    [
+        param(
+            {},
+            {
+                "content_type": "auto",
+                "summary_length": "brief",
+                "summary_type": "bullets",
+            },
+            id="default params",
+        ),
+        param(
+            {"content_type": "informative"},
+            {
+                "content_type": "informative",
+                "summary_length": "brief",
+                "summary_type": "bullets",
+            },
+            id="set content_type param",
+        ),
+        param(
+            {"summary_length": "detailed"},
+            {
+                "content_type": "auto",
+                "summary_length": "detailed",
+                "summary_type": "bullets",
+            },
+            id="set summary_length param",
+        ),
+        param(
+            {"summary_type": "paragraphs"},
+            {
+                "content_type": "auto",
+                "summary_length": "brief",
+                "summary_type": "paragraphs",
+            },
+            id="set summary_type param",
+        ),
+        param(
+            {
+                "content_type": "auto",
+                "summary_length": "brief",
+                "summary_type": "bullets",
+            },
+            {
+                "content_type": "auto",
+                "summary_length": "brief",
+                "summary_type": "bullets",
+            },
+            id="set all params",
+        ),
+    ],
+)
+def test_summarization_config(params, want):
+    summarization_config = models.SummarizationConfig(**params)
+    assert asdict(summarization_config) == want
+
+
+@mark.parametrize(
+    "params, want",
+    [
+        param(
+            {},
+            {"topics": None},
+            id="default params",
+        ),
+        param(
+            {"topics": ["topic1", "topic2"]},
+            {"topics": ["topic1", "topic2"]},
+            id="set topics param",
+        ),
+    ],
+)
+def test_topic_detection_config(params, want):
+    topic_detection_config = models.TopicDetectionConfig(**params)
+    assert asdict(topic_detection_config) == want
+
+
+@mark.parametrize(
+    "params, want",
+    [
+        param(
+            {"url": "example.com"},
+            {
+                "url": "example.com",
+                "contents": None,
+                "method": "post",
+                "auth_headers": None,
+            },
+        ),
+        param(
+            {
+                "url": "example.com",
+                "contents": ["transcript.txt"],
+                "method": "put",
+                "auth_headers": ["Authorization", "Bearer token"],
+            },
+            {
+                "url": "example.com",
+                "contents": ["transcript.txt"],
+                "method": "put",
+                "auth_headers": ["Authorization", "Bearer token"],
+            },
+        ),
+    ],
+)
+def test_notification_config(params, want):
+    notification_config = models.NotificationConfig(**params)
+    assert asdict(notification_config) == want